feat(warp): don't unnecessarily save interactions PE-4967
Adds a simple check that compares the number of interactions already
saved for the given wallet and epoch with the number that need to be
saved. If the number saved is greater than or equal to the expected
count, no new interactions are saved. This assumes that the same report
is used passed in each time. Currently, the contents of the interactions
are not checked. In the future, we may want to compare contract state or
extract the gateways already reported from the GraphQL results and
compare that to the expected list.